vector_service/requirements.txt
# VectorService Dependencies
# Embedding model, Pinecone, RabbitMQ consumer, HTTP client

fastapi==0.104.1
uvicorn==0.24.0
pinecone-client==2.2.4
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
pika==1.3.2
httpx==0.25.1
requests==2.31.0
numpy==1.26.2
pydantic==2.5.0
python-dotenv==1.0.0

